+++ force_wfmanager/gui/run.py
@@ -3,6 +3,7 @@ import logging
 
 from envisage.core_plugin import CorePlugin
 from envisage.ui.tasks.tasks_plugin import TasksPlugin
+from force_wfmanager.wfmanager_plugin import WfManagerPlugin
 
 from stevedore import extension
 from stevedore.exception import NoMatches
@@ -37,7 +38,8 @@ def main(workflow_file=None, debug=False):
                             level=logging.DEBUG)
     log = logging.getLogger(__name__)
 
-    plugins = [CorePlugin(), TasksPlugin(), FactoryRegistryPlugin()]
+    plugins = [CorePlugin(), TasksPlugin(), FactoryRegistryPlugin(),
+               WfManagerPlugin(workflow_file=workflow_file)]
 
     mgr = extension.ExtensionManager(
         namespace='force.bdss.extensions',
@@ -53,5 +55,5 @@ def main(workflow_file=None, debug=False):
     except NoMatches:
         log.info("No extensions found")
 
-    wfmanager = WfManager(plugins=plugins, workflow_file=workflow_file)
+    wfmanager = WfManager(plugins=plugins)
     wfmanager.run()
