+++ force_wfmanager/left_side_pane/results_pane.py
@@ -39,6 +39,10 @@ class ResultsPane(TraitsDockPane):
         UItem('result_table', style='custom'),
     ))
 
+    def __init__(self, analysis_model, *args, **kwargs):
+        self.analysis_model=analysis_model
+        super(ResultsPane, self).__init__(*args, **kwargs)
+
     def _result_table_default(self):
         return ResultTable(
             analysis_model=self.analysis_model
