+++ force_wfmanager/left_side_pane/tree_pane.py
@@ -61,6 +61,7 @@ class TreePane(TraitsDockPane):
             factory_registry=self.factory_registry,
             model=self.workflow_m
         )
+        print(self.factory_registry)
         self.run_enabled = wf_tree.workflow_mv.valid
         return wf_tree
