+++ force_wfmanager/plugins/ui_notification/ui_notification_hooks_manager.py
@@ -10,49 +10,49 @@ class UINotificationHooksManager(BaseUIHooksManager):
     When the execution ends, the added information is removed.
     """
 
-    def before_execution(self, app):
+    def before_execution(self, task):
         """Sets up the ui_notification notification listener, which
         listens to the ports for the ZMQ server for incoming data
         from the bdss.
 
         Parameters
         ----------
-        app:
-            The top-level WfManager instance containing the workflow
+        task:
+            The WfManagerResultsTask instance containing the workflow
             model and the zmq server.
         """
-        model = app.workflow_m
+        model = task.workflow_m
         notification_model = None
         for listener_model in model.notification_listeners:
             if isinstance(listener_model, UINotificationModel):
                 notification_model = listener_model
 
         if notification_model is None:
-            registry = app.factory_registry
+            registry = task.factory_registry
             nl_factory = registry.notification_listener_factory_by_id(
                 factory_id(self.factory.plugin.id, "ui_notification")
             )
             notification_model = nl_factory.create_model()
             model.notification_listeners.append(notification_model)
 
-        pub_port, sync_port = app.zmq_server.ports
+        pub_port, sync_port = task.zmq_server.ports
         notification_model.sync_url = (
                 "tcp://127.0.0.1:"+str(sync_port))
         notification_model.pub_url = (
                 "tcp://127.0.0.1:"+str(pub_port))
         notification_model.identifier = ""
 
-    def after_execution(self, app):
+    def after_execution(self, task):
         """Removes the ui_notification notification listener, as
         the bdss has finished calculating.
 
         Parameters
         ----------
-        app:
-            The top-level WfManager instance containing the workflow
+        task:
+            The WfManagerResultsTask instance instance containing the workflow
             model and the zmq server.
         """
-        model = app.workflow_m
+        model = task.workflow_m
         notification_model = None
         for listener_model in model.notification_listeners:
             if isinstance(listener_model, UINotificationModel):
